requirements.txt
backcall==0.1.0
cached-property==1.5.1
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
idna==2.8
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.16.4
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
python-bugzilla==2.2.0
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
